modif entity showRate : DiscountRate en float

+++ src/Entity/ShowRate.php
@@ -25,7 +25,7 @@ class ShowRate
 
     /**
      * Display the discounted Rate for the show if this offer has been chosen.
-     * @ORM\Column(type="integer", nullable=true)
+     * @ORM\Column(type="float", nullable=true)
      */
     private $discountedRate;
 
@@ -52,12 +52,12 @@ class ShowRate
         return $this;
     }
 
-    public function getDiscountedRate(): ?int
+    public function getDiscountedRate(): ?float
     {
         return $this->discountedRate;
     }
 
-    public function setDiscountedRate(?int $discountedRate): self
+    public function setDiscountedRate(?float $discountedRate): self
     {
         $this->discountedRate = $discountedRate;
